% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{as.data.frame.dce}
\alias{as.data.frame.dce}
\title{Turn dce object into data frame}
\usage{
\method{as.data.frame}{dce}(x, row.names = NULL, optional = FALSE, ...)
}
\value{
data frame containing the dce output
}
\description{
Turn dce object into data frame
}
\examples{
dag <- create_random_DAG(30, 0.2)
X_wt <- simulate_data(dag)
dag_mt <- resample_edge_weights(dag)
X_mt <- simulate_data(dag_mt)
dce_list <- dce(dag, X_wt, X_mt)
}
