% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_adjmat}
\alias{as_adjmat}
\title{graphNEL with 0 edge weights to proper adjacency matrix}
\usage{
as_adjmat(g)
}
\arguments{
\item{g}{graphNEL object}
}
\description{
graphNEL with 0 edge weights to proper adjacency matrix
}
