% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_causal_effects}
\alias{compute_causal_effects}
\title{Compute the causal effect of all nodes on al other nodes of a given DAG}
\usage{
compute_causal_effects(graph, df.expr)
}
\arguments{
\item{graph}{DAG as a graphNEL object}

\item{df.expr}{genes expression profiles with rows as observations and
columns as variables (nodes)}
}
\value{
vector of causal effects
}
\description{
This function takes a DAG and gene expression data as input and computes
the causal effect of all nodes on all other nodes in the DAG.
}
\examples{
dag <- matrix(c(0,0,0,1,0,0,0,1,0), 3)
colnames(dag) <- rownames(dag) <- seq_len(3)
dag <- as(dag, "graphNEL")
d <- matrix(rnorm(100*3), 100)
colnames(d) <- seq_len(3)
compute_causal_effects(dag, d)
}
\author{
Kim Jablonski
}
