% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_differential_causal_effects}
\alias{compute_differential_causal_effects}
\title{Compute the differential causal effects}
\usage{
compute_differential_causal_effects(
  graph.ctrl,
  df.expr.ctrl,
  graph.mut,
  df.expr.mut,
  method = "full",
  bootstrap = FALSE,
  runs = 100,
  replace = FALSE,
  frac = 0.5,
  bootMethod = "diff",
  errDist = "nbinom",
  theta = NULL,
  partial = FALSE,
  link.log.base = exp(1),
  ...
)
}
\arguments{
\item{graph.ctrl}{DAG as a graphNEL object of the control population}

\item{df.expr.ctrl}{Gene expression profiles with rows as observations and
columns as variables (nodes) of the control population}

\item{graph.mut}{DAG as a graphNEL object of the tumor population}

\item{df.expr.mut}{genes expression profiles with rows as observations and
columns as variables (nodes) of the tumor population}

\item{method}{Either "full" for the full linear model using all
samples together to directly compute the differential causal effect
or "part" for two linear models on both conditions to first compute
the causal effect and then the difference}

\item{bootstrap}{if TRUE, uses a bootstrap or subsample routine to
compute the effects}

\item{runs}{bootstrap/subsampling runs}

\item{replace}{if TRUE, classic bootstrap, if FALSE subsampling without
replacement}

\item{frac}{the fraction of the data to sample, can either be one value
or a vector of two for different fractions for control and tumor,
respectively populations}

\item{bootMethod}{if "diff" (default) bootstraps the differential causal
effects, if "cause", bootstraps on the causal effects
(not possible for "full")}

\item{theta}{prior estimated theta value}

\item{partial}{if TRUE only computes the partial causal effects on
the edges, else computes the total causal effect}

\item{link.log.base}{base of logarithm of link}

\item{...}{additional parameters for glm2::glm2}

\item{conf}{if TRUE accounts for confounders}
}
\value{
vector of causal effects
}
\description{
This function takes a DAG and gene expression data of two distinct types
of populations (e.g. tumor versus normal) as input and computes
the differential causal effects of all nodes on all other nodes in the DAG
between the tow populations of samples.
}
\examples{
graph.wt <- as(matrix(c(0,0,0,1,0,0,0,1,0), 3), "graphNEL")
graph.mt <- resample_edge_weights(graph.wt)
X.wt <- simulate_data(graph.wt)
X.mt <- simulate_data(graph.mt)
compute_differential_causal_effects(graph.wt, X.wt, graph.mt, X.mt)
}
\author{
Kim Jablonski & Martin Pirkl
}
