% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_enrichment}
\alias{compute_enrichment}
\title{Compute pathway enrichment}
\usage{
compute_enrichment(
  graph,
  X.wt,
  X.mt,
  statistic = function(x) {     sum(abs(x)) },
  permutation_count = 100,
  pvalue.method = "hmp",
  theta = NULL,
  partial = 1,
  dce = NULL
)
}
\arguments{
\item{graph}{DAG as a graphNEL object}

\item{X.wt}{Expression values of wild type (if dce=NULL)}

\item{X.mt}{Expression values of mutant (if dce=NULL)}

\item{statistic}{Statistic to compute}

\item{permutation_count}{How many permutations to do}

\item{theta}{prior estimated theta value}

\item{partial}{if TRUE only computes the partial causal effects on}

\item{dce}{optional dce object
the edges, else computes the total causal effect}
}
\value{
Enrichment p-value
}
\description{
This function computes a p-value.
}
\examples{
graph.wt <- as(matrix(c(0,0,0,1,0,0,0,1,0), 3), "graphNEL")
graph.mt <- resample_edge_weights(graph.wt)
X.wt <- simulate_data(graph.wt)
X.mt <- simulate_data(graph.mt)
compute_enrichment(graph.wt, X.wt, X.mt)
}
\author{
Hinz und Kunz
}
