% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_permutations}
\alias{compute_permutations}
\title{Compute statistic for permutation test}
\usage{
compute_permutations(
  normal,
  dn,
  tumor,
  dt,
  runs = 10,
  statistic = function(x) return(sum(abs(x))),
  ...
)
}
\arguments{
\item{runs}{number of permutations}

\item{statistic}{function, which computes a single valued
statistic for each run}

\item{...}{additional parameters for compute_differential_effects}

\item{graph.ctrl}{DAG as a graphNEL object of the control population}

\item{df.expr.ctrl}{Gene expression profiles with rows as observations and
columns as variables (nodes) of the control population}

\item{graph.mut}{DAG as a graphNEL object of the tumor population}

\item{df.expr.mut}{genes expression profiles with rows as observations and
columns as variables (nodes) of the tumor populationt}
}
\value{
numeric vector with permutation statistics of length runs
}
\description{
This function takes the same parameters as
compute_differential_effects in addition to
control parameters for the permutations.
}
\examples{
1
}
\author{
Martin Pirkl
}
