% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{create_graph_from_dataframe}
\alias{create_graph_from_dataframe}
\title{Create graph from dataframe}
\usage{
create_graph_from_dataframe(
  df_graph,
  edge_weight = function() runif(1, 0.5, 2)
)
}
\arguments{
\item{df_graph}{Dataframe with source and target columns}

\item{edge_weight}{Edge weight (function)}
}
\description{
Create graph from dataframe
}
