% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_random_DAG}
\alias{create_random_DAG}
\title{Create random DAG (topologically ordered)}
\usage{
create_random_DAG(
  n,
  prob,
  lB = -1,
  uB = 1,
  node_labels = paste0("n", as.character(seq_len(n)))
)
}
\arguments{
\item{n}{Number of nodes}

\item{prob}{Probability of creating an edge}

\item{lB}{Lower bound for edge weights}

\item{uB}{Upper bound for edge weights}

\item{node_labels}{Node labels}
}
\value{
graph
}
\description{
Creates a DAG according to given parameters.
}
\examples{
dag <- create_random_DAG(30, 0.2)
}
\author{
Martin Pirkl
}
