% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_implementation.R
\name{dce}
\alias{dce}
\title{dce - main function}
\usage{
dce(
  graph,
  df.expr.wt,
  df.expr.mt,
  solver = "glm2",
  solver.args = list(method = "glm.dce.fit"),
  adjustment.type = "parents",
  p.method = "meanp",
  verbose = FALSE
)
}
\arguments{
\item{graph}{valid object defining a directed acyclic graph}

\item{df.expr.wt}{data frame with wild type expression values}

\item{df.expr.mt}{data from with mutation type expression values}

\item{solver}{character with name of solver function}

\item{solver.args}{additional arguments for the solver function}

\item{adjustment.type}{character string for the method to define
the adjustment set Z for the regression}

\item{p.method}{character string. "hmp" for harmonic mean
or any method from package 'metap', e.g., "meanp" or "sump"}

\item{verbose}{logical for verbose output}
}
\description{
Main function to compute differential causal effects and its enrichment
}
