% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_adjustment_set}
\alias{get_adjustment_set}
\title{Get adjustment set on graph given two nodes}
\usage{
get_adjustment_set(
  graph,
  x,
  y,
  adjustment_type = "parents",
  effect_type = "total"
)
}
\arguments{
\item{graph}{Topology to use}

\item{x}{Source node}

\item{y}{target node}

\item{adjustment_type}{Which adjustment method to use}

\item{effect_type}{Which effect to compute}
}
\description{
Get adjustment set on graph given two nodes
}
