% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_databases.R
\name{get_pathway_info}
\alias{get_pathway_info}
\title{Dataframe containing meta-information of pathways in database}
\usage{
get_pathway_info(
  query_species = "hsapiens",
  database_list = NULL,
  include_network_statistics = FALSE
)
}
\arguments{
\item{query_species}{For which species}

\item{database_list}{Which databases to query. Query all if `NULL`.}

\item{include_network_statistics}{Compute some useful statistics
per pathway. Takes longer!}
}
\description{
Dataframe containing meta-information of pathways in database
}
