% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_databases.R
\name{get_pathways}
\alias{get_pathways}
\title{Easy pathway network access}
\usage{
get_pathways(
  query_species = "hsapiens",
  database_list = NULL,
  remove_empty_pathways = TRUE,
  pathway_list = NULL
)
}
\arguments{
\item{query_species}{For which species}

\item{database_list}{Which databases to query. Query all if `NULL`.}

\item{remove_empty_pathways}{Discard pathways without nodes}

\item{pathway_list}{List mapping database name to
vector of pathway names to download}
}
\value{
list of pathways
}
\description{
Easy pathway network access
}
\examples{
pathways <- get_pathways(
  pathway_list = list(kegg = c(
    "Protein processing in endoplasmic reticulum"
  ))
)
plot_network(as(pathways[[1]]$graph, "matrix"))
}
