% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_prediction_counts}
\alias{get_prediction_counts}
\title{Compute true positive/... counts}
\usage{
get_prediction_counts(truth, inferred, cutoff = 0.5)
}
\arguments{
\item{truth}{Ground truth}

\item{inferred}{Computed results}

\item{cutoff}{Threshold for classification}
}
\value{
data.frame
}
\description{
Useful for performance evaluations
}
\examples{
get_prediction_counts(c(1,0), c(1,1))
}
\author{
Hans Wurst
}
