% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glm.dce.fit}
\alias{glm.dce.fit}
\title{modified glm.fit function}
\usage{
glm.dce.fit(
  x,
  y,
  weights = rep(1, nobs),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = rep(0, nobs),
  family = gaussian(),
  control = list(),
  intercept = TRUE,
  singular.ok = TRUE
)
}
\description{
modified glm.fit function
}
