% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_implementation.R
\name{glm.dce.nb.fit}
\alias{glm.dce.nb.fit}
\title{Extra robust model fitting}
\usage{
glm.dce.nb.fit(...)
}
\description{
`glm.nb` uses two families: poisson (initial theta fit) and negative binomial (final fit).
The solver can fail if mu becomes negative
}
