% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_glm.R
\name{glm.nb.rob}
\alias{glm.nb.rob}
\title{Modified glm.nb function from package MASS}
\usage{
glm.nb.rob(
  formula,
  data,
  weights,
  subset,
  na.action,
  start = NULL,
  etastart,
  mustart,
  control = glm.control(...),
  method = "glm.fit",
  model = TRUE,
  x = FALSE,
  y = TRUE,
  contrasts = NULL,
  ...,
  init.theta,
  link = log
)
}
\description{
fixes a bug, if theta estimation breaks
see ?MASS::glm.nb for argument values
}
