% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{glm_solver}
\alias{glm_solver}
\title{Generic solver function}
\usage{
glm_solver(form, df, solver, solver_args)
}
\arguments{
\item{form}{Formula describing model}

\item{df}{Data used to fit model}

\item{solver}{Which solver to use}

\item{solver_args}{Additional arguments for solver}
}
\description{
Generic solver function
}
