% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{graph2df}
\alias{graph2df}
\title{Convert graph object to dataframe with source and target columns}
\usage{
graph2df(graph)
}
\arguments{
\item{graph}{Network}
}
\description{
Convert graph object to dataframe with source and target columns
}
