% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{graph_union}
\alias{graph_union}
\title{Create union of multiple graphs}
\usage{
graph_union(graph_list)
}
\arguments{
\item{graph_list}{List of graphs}
}
\description{
Create union of multiple graphs
}
