% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pcor}
\alias{pcor}
\title{Partial correlation}
\usage{
pcor(x, g = NULL, adjustment_type = "parents", ...)
}
\arguments{
\item{x}{matrix}

\item{g}{related graph as adjacency matrix (optional)}

\item{adjustment_type}{character string for the method
to define the adjustment set Z for the regression}

\item{...}{additional arguments for function 'cor'}
}
\value{
matrix of partial correlations
}
\description{
Robust partial correlation of column variables of a
numeric matrix
}
\examples{
x <- matrix(rnorm(100),10,10)
pcor(x)
}
