% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pcor_perm}
\alias{pcor_perm}
\title{permutation test for (partial) correlation on non-Gaussian data}
\usage{
pcor_perm(x, y, iter = 1000, fun = pcor, method = "spearman")
}
\arguments{
\item{x}{wild type data set}

\item{y}{mutant data set}

\item{fun}{function to compute the statistic, e.g., cor or pcor}

\item{i}{number of iterations (permutations)}
}
\description{
permutation test for (partial) correlation on non-Gaussian data
}
