% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permutation_test}
\alias{permutation_test}
\title{Permutation test for (partial) correlation on non-Gaussian data}
\usage{
permutation_test(x, y, iter = 1000, fun = pcor, ...)
}
\arguments{
\item{x}{wild type data set}

\item{y}{mutant data set}

\item{iter}{number of iterations (permutations)}

\item{fun}{function to compute the statistic, e.g., cor or pcor}

\item{...}{additional arguments for function 'fun'}
}
\value{
matrix of p-values
}
\description{
Computes the significance of (partial) correlation based
on permutations of the observations
}
\examples{
x <- matrix(rnorm(100),10,10)
y <- matrix(rnorm(100),10,10)
permutation_test(x,y,iter=10)
}
