% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{plot.dceSim}
\alias{plot.dceSim}
\title{Plot simulation study}
\usage{
\method{plot}{dceSim}(
  x,
  col = seq_len(4),
  border = col,
  showMeth = seq_len(4),
  showFeat = 1,
  methNames = NULL,
  ...
)
}
\arguments{
\item{x}{dceSim object}

\item{col}{`col` argument passed to `boxplot`}

\item{border}{`border` argument passed to `boxplot`}

\item{showMeth}{which method results to plot}

\item{showFeat}{which features to plot}

\item{methNames}{method plot labels}

\item{...}{additional arguments for boxplot}
}
\value{
plot
}
\description{
Takes a dceSim object and produces a figure.
}
\author{
Martin Pirkl
}
