% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotDce}
\alias{plotDce}
\title{Plot dce}
\usage{
plotDce(
  x,
  type = "graph",
  log = FALSE,
  scalefac = NULL,
  edgecol = NULL,
  edgewidth = NULL,
  genelabels = NULL,
  ...
)
}
\arguments{
\item{x}{object of class dce}

\item{type}{"graph" or "heatmap"}

\item{log}{if TRUE, applies log to dces}

\item{scalefac}{scalar for scaling the dces}

\item{edgecol}{optional}

\item{edgewidth}{otpional}

\item{genelabels}{named vector with names like the original names
used in the dce object and values as subtitute names}

\item{...}{more arguments for mnem::plotDnf or epiNEM::HeatmapOP}
}
\description{
Wrapper function for mnem::plotDnf and epiNEM::HeatmapOP
}
