% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{propagate_gene_edges}
\alias{propagate_gene_edges}
\title{Remove non-gene nodes from pathway and reconnect nodes}
\usage{
propagate_gene_edges(graph)
}
\arguments{
\item{graph}{Biological pathway}
}
\description{
Remove non-gene nodes from pathway and reconnect nodes
}
