% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resample_edge_weights}
\alias{resample_edge_weights}
\title{Resample network edge weights}
\usage{
resample_edge_weights(g, tp = 0.5, mineff = 1, maxeff = 2, method = "runif")
}
\arguments{
\item{g}{original graph}

\item{tp}{fraction of edge weights which will be modified}

\item{mineff}{minimal differential effect size}

\item{maxeff}{maximum effect effect size or standard deviation,
if method is "gauss"}

\item{method}{method for drawing the differential for the causal effects.
Can be "runif", "exp" or "gauss".}
}
\value{
graph with new edge weights
}
\description{
Takes a graph and modifies edge weights.
}
\examples{
graph.wt <- as(matrix(c(0,0,0,1,0,0,0,1,0), 3), "graphNEL")
graph.mt <- resample_edge_weights(graph.wt)
}
\author{
Martin Pirkl
}
