% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simDce}
\alias{simDce}
\title{Simulation study}
\usage{
simDce(
  nodes = 5,
  samples = c(10, 10),
  simruns = 10,
  mu = 0,
  sd = 1,
  effRange = c(-1, 0, 0, 1),
  truePos = 1,
  perturb = 0,
  cormeth = "p",
  prob = "runif",
  bootstrap = 0,
  verbose = FALSE,
  test = 0,
  enriched = 1,
  theta = NULL,
  partial = TRUE,
  ...
)
}
\arguments{
\item{nodes}{numebr of genes}

\item{samples}{vector of length two for sample number of both conditions}

\item{simruns}{simulations runs}

\item{mu}{mean expression}

\item{sd}{standard deviation}

\item{effRange}{vector of length four with lowest, second lowest, second
highest and highest possible effect the effects are uniformly drawn from}

\item{truePos}{the fraction of differential effects; truePos=1 means all
are potentially differential (see effRange); if truePos=0.5, 50% of
differential effects are set to 0.}

\item{perturb}{positive or negative frantion of edges to be added or
removed, respectively}

\item{cormeth}{method for the correlation accuracy (see ?cor)}

\item{prob}{edge probability; either probability or "runif" to draw a
probability in each run}

\item{bootstrap}{can be either 0 (default) or have the first argument as
the number of runs and include one or more
of the following: "basic", "full"}

\item{verbose}{verbose output, if TRUE}

\item{test}{if greater than 0, tests the pathway for enrichment with
test permutation runs}

\item{enriched}{fraction of runs with enriched pathways}

\item{theta}{theta for negative binomial model}

\item{partial}{only compute the edge weights}

\item{...}{additional parameters for compute_differential_causal_effects}
}
\value{
accuracy for several different methods and statistics
for the ground truth as a list of two arrays
}
\description{
This function takes several parameters to define a simulation study.
}
\examples{
1
}
\author{
Martin Pirkl
}
