% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data}
\usage{
simulate_data(
  graph,
  n = 100,
  dist.mean = 1000,
  dist.dispersion = 100,
  link = negative.binomial.special()$linkfun
)
}
\arguments{
\item{n}{Number of samples}

\item{dist.mean}{distribution mean as numeric}

\item{dist.dispersion}{distribution dispersion
(actually dispersion^-1) as a scalar}

\item{dag}{Graph to simulate on}
}
\value{
graph
}
\description{
Generate data for given DAG.
}
\examples{
dag <- create_random_DAG(30, 0.2)
X <- simulate_data(dag)
}
