% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{simulate_data_old}
\alias{simulate_data_old}
\title{Simulate data}
\usage{
simulate_data_old(
  dag,
  n = 100,
  errDist = c("normal", "cauchy", "t4", "mix", "mixt3", "mixN100", "nbinom"),
  normpars = c(0, 1),
  mix = 0.1,
  errMat = NULL,
  back.compatible = FALSE,
  use.node.names = !back.compatible
)
}
\arguments{
\item{dag}{Graph to simulate on}

\item{n}{Number of samples}

\item{errDist}{Noise distribution}

\item{normpars}{Noise parameters}

\item{mix}{Another noise parameter}

\item{errMat}{Initial error matrix}

\item{back.compatible}{Legacy compatibility}

\item{use.node.names}{Keep node names}
}
\value{
graph
}
\description{
Generate data for given DAG.
}
\examples{
dag <- create_random_DAG(30, 0.2)
X <- simulate_data(dag)
}
\author{
Martin Pirkl
}
