% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.glmmle}
\alias{summary.glmmle}
\title{Estimate p-values}
\usage{
\method{summary}{glmmle}(x)
}
\arguments{
\item{x}{object of class "glmmle"}
}
\value{
data.frame with coefficients and p-values
}
\description{
Function to compute p-values based on regression coefficients
}
\author{
Martin Pirkl
}
