% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.rlm_dce}
\alias{summary.rlm_dce}
\title{summary for rlm_dce}
\usage{
\method{summary}{rlm_dce}(object, ...)
}
\arguments{
\item{object}{object of class 'rlm_dce'}

\item{...}{see ?MASS::summary.rlm}
}
\description{
summary for rlm_dce
}
