% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{symlog}
\alias{symlog}
\title{Linear transform if under logarithmic transform is over threshold}
\usage{
symlog(x, base = 10, threshold = 1)
}
\arguments{
\item{x}{Value to transform}

\item{base}{Base of logarithm}

\item{threshold}{Linearity threshold}
}
\description{
Linear transform if under logarithmic transform is over threshold
}
