% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_mle.R
\name{theta.ml.rob}
\alias{theta.ml.rob}
\title{Modified theta.ml function from package MASS}
\usage{
theta.ml.rob(
  y,
  mu,
  n = sum(weights),
  weights,
  limit = 10,
  eps = .Machine$double.eps^0.25,
  trace = FALSE
)
}
\description{
fixes a bug, if theta estimation breaks
see ?MASS::theta.ml for argument values
}
