% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trueEffects}
\alias{trueEffects}
\title{Compute the true casual effects of a simulated dag}
\usage{
trueEffects(g, partial = FALSE)
}
\arguments{
\item{g}{graphNEL object}

\item{partial}{if FALSE computes the total causal effects and not just
the partial edge effects}
}
\value{
matrix of causal effects
}
\description{
This function takes a DAG with edgeweights as input and computes
the causal effects of all nodes on all direct and indirect children in the
DAG. Alternatively see pcalg::causalEffect for pairwise computation.
}
\examples{
graph.wt <- as(matrix(c(0,0,0,1,0,0,0,1,0), 3), "graphNEL")
trueEffects(graph.wt)
}
\author{
Martin Pirkl
}
