% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_adjmat}
\alias{as_adjmat}
\title{graph to adjacency}
\usage{
as_adjmat(g)
}
\arguments{
\item{g}{graphNEL object}
}
\value{
graph as adjacency matrix
}
\description{
From graphNEL with 0 edge weights to proper adjacency matrix
}
\examples{
dag <- create_random_DAG(30, 0.2)
adj <- as_adjmat(dag)
}
