% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_random_DAG}
\alias{create_random_DAG}
\title{Create random DAG (topologically ordered)}
\usage{
create_random_DAG(
  node_num,
  prob,
  eff_min = -1,
  eff_max = 1,
  node_labels = paste0("n", as.character(seq_len(node_num))),
  max_par = 3
)
}
\arguments{
\item{node_num}{Number of nodes}

\item{prob}{Probability of creating an edge}

\item{eff_min}{Lower bound for edge weights}

\item{eff_max}{Upper bound for edge weights}

\item{node_labels}{Node labels}

\item{max_par}{Maximal number of parents}
}
\value{
graph
}
\description{
Creates a DAG according to given parameters.
}
\examples{
dag <- create_random_DAG(30, 0.2)
}
\author{
Martin Pirkl
}
