% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dce}
\alias{dce}
\title{dce - main function}
\usage{
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = "glm.dce.fit"),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{graph}{valid object defining a directed acyclic graph}

\item{df_expr_wt}{data frame with wild type expression values}

\item{df_expr_mt}{data from with mutation type expression values}

\item{solver}{character with name of solver function}

\item{solver_args}{additional arguments for the solver function}

\item{adjustment_type}{character string for the method to define
the adjustment set Z for the regression}

\item{effect_type}{method of computing causal effects}

\item{p_method}{character string. "mean", "sum" for standard summary
functions, "hmp" for harmonic mean, "test" for the selfcontained
test of package 'CombinePValue' or any method from package 'metap',
e.g., "meanp" or "sump".}

\item{test}{either "wald" for testing significance with the
wald test or "lr" for using a likelihood ratio test}

\item{lib_size}{either a numeric vector of the same length as the
sum of wild type and mutant samples or a logical. If TRUE, it is
recommended that both data sets include not only the genes
included in the graph but all genes available in the original data set.}

\item{deconfounding}{indicates whether adjustment against latent 
confounding is used. If FALSE, no adjustment is used, if TRUE it adjusts
for confounding by automatically estimating the number of latent
confounders. The estimated number of latent confounders can be chosen
manually by setting this variable to some number.}

\item{verbose}{logical for verbose output}
}
\value{
list of matrices with dces and corresponding p-value
}
\description{
Main function to compute differential causal effects and its enrichment
}
\examples{
dag <- create_random_DAG(30, 0.2)
X.wt <- simulate_data(dag)
dag.mt <- resample_edge_weights(dag)
X.mt <- simulate_data(dag)
dce(dag,X.wt,Xmt)
}
