% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pathway_statistics}
\alias{df_pathway_statistics}
\title{Biological pathway information.}
\format{
A data frame with pathway statistics
\describe{
  \item{database}{Pathway database}
  ...
}
}
\usage{
df_pathway_statistics
}
\description{
A dataset containing pathway statistics.
}
\keyword{datasets}
