% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_latent_confounder_count}
\alias{estimate_latent_confounder_count}
\title{Estimate number of latent confounders.}
\usage{
estimate_latent_confounder_count(X)
}
\description{
This function looks at the knee point of a scree plot.
}
