% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_glm.R
\name{glm.dce.fit}
\alias{glm.dce.fit}
\title{Identity robust glm fit.}
\usage{
glm.dce.fit(
  x,
  y,
  weights = rep(1, nobs),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = rep(0, nobs),
  family = gaussian(),
  control = list(),
  intercept = TRUE,
  singular.ok = TRUE
)
}
\description{
A modified glm.fit function which prevents negative mean
values in case of the identity link function. Convergence
is not guaranteed!
}
