% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pcor}
\alias{pcor}
\title{Partial correlation}
\usage{
pcor(x, method = "spearman")
}
\arguments{
\item{x}{matrix}
}
\value{
matrix of partial correlations
}
\description{
Robust partial correlation of column variables of a 
numeric matrix
}
\examples{
x <- matrix(rnorm(100),10,10)
pcor(x)
}
