% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.dce}
\alias{plot.dce}
\title{Plot dce object}
\usage{
\method{plot}{dce}(x, ...)
}
\arguments{
\item{x}{dce object}

\item{...}{Parameters passed to dce::plot_network}
}
\value{
plot of dag and dces
}
\description{
This function takes a differential causal effects object and plots
the dag with the dces
}
\examples{
dag <- create_random_DAG(30, 0.2)
X.wt <- simulate_data(dag)
dag.mt <- resample_edge_weights(dag)
X.mt <- simulate_data(dag)
dce.list <- dce(dag,X.wt,X.mt)
plot(dce.list)
}
\author{
Martin Pirkl, Kim Philipp Jablonski
}
