% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.dce}
\alias{plot.dce}
\title{Plot dce object}
\usage{
\method{plot}{dce}(x, ...)
}
\arguments{
\item{x}{dce object}

\item{...}{Parameters passed to dce::plot_network}
}
\value{
plot of dag and dces
}
\description{
This function takes a differential causal effects object and plots
the dag with the dces
}
\author{
Martin Pirkl, Kim Philipp Jablonski
}
