% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.dce}
\alias{plot.dce}
\title{Plot dce object}
\usage{
\method{plot}{dce}(x, nodename.map = NULL, edge.colorscale.limits = NULL, ...)
}
\arguments{
\item{x}{dce object}

\item{nodename.map}{node names}

\item{edge.colorscale.limits}{Limits for scale_edge_color_gradient2 (should contain 0). Useful to make plot comparable to others}

\item{...}{additional parameters}
}
\value{
plot of dag and dces
}
\description{
This function takes a differential causal effects object and plots
the dag with the dces
}
\author{
Martin Pirkl, Kim Philipp Jablonski
}
