% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resample_edge_weights}
\alias{resample_edge_weights}
\title{Resample network edge weights}
\usage{
resample_edge_weights(g, lB = -1, uB = 1, tp = 1)
}
\arguments{
\item{g}{original graph}

\item{lB}{lower bound for new edge weights}

\item{uB}{upper bound for new edge weights}

\item{tp}{fraction of edge weights which will be modified}
}
\value{
graph with new edge weights
}
\description{
Takes a graph and modifies edge weights.
}
\examples{
graph.wt <- as(matrix(c(0,0,0,1,0,0,0,1,0), 3), "graphNEL")
graph.mt <- resample_edge_weights(graph.wt)
}
\author{
Martin Pirkl
}
