% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data}
\usage{
simulate_data(
  graph,
  n = 100,
  dist_mean = 1000,
  dist_dispersion = 100,
  link = negative.binomial.special()$linkfun,
  pop_size = 100,
  latent = 0
)
}
\arguments{
\item{n}{Number of samples}

\item{dist_mean}{distribution mean as numeric}

\item{dist_dispersion}{distribution dispersion
(actually dispersion^-1) as a scalar}

\item{dag}{Graph to simulate on}
}
\value{
graph
}
\description{
Generate data for given DAG.
}
\examples{
dag <- create_random_DAG(30, 0.2)
X <- simulate_data(dag)
}
