% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data}
\usage{
simulate_data(
  graph,
  n = 100,
  dist_mean = 1000,
  dist_dispersion = 100,
  link = negative.binomial.special()$linkfun,
  pop_size = 0,
  latent = 0
)
}
\arguments{
\item{graph}{Graph to simulate on}

\item{n}{Number of samples}

\item{dist_mean}{distribution mean as numeric}

\item{dist_dispersion}{distribution dispersion
(actually dispersion^-1) as a scalar}

\item{link}{special link function for the negative binomial
distribution}

\item{pop_size}{numeric for the population size, e.g., pop_size=1000 adds
1000-n random genes not in the graph}

\item{latent}{number of latent variables}
}
\value{
graph
}
\description{
Generate data for given DAG.
}
\examples{
dag <- create_random_DAG(30, 0.2)
X <- simulate_data(dag)
}
