% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{topologically_ordering}
\alias{topologically_ordering}
\title{Order rows/columns topologically}
\usage{
topologically_ordering(adja_mat, alt = FALSE)
}
\arguments{
\item{adja_mat}{Adjacency matrix of network}

\item{alt}{Use igraph implementation}
}
\description{
Order rows/columns topologically
}
